requirements.txt
# --- Core Framework ---
Django==5.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
# --- Database & Async ---
psycopg[binary]==3.1.18
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# --- API Documentation ---
drf-spectacular==0.27.1

# --- Utilities ---
django-environ==0.11.2

# --- Testing & Quality (Dev only) ---
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
ruff==0.3.0
pytest-cov==4.1.0